Changed to let the Supervisor's saver to save the variables in the end
Line 294: If we use `saver.save` instead of `sv.saver.save`, then the excluded variables that were not restored will not be saved in the checkpoint file as well. This causes an issue if you end the training early before the default 10 minutes that prompts the Supervisor's saver to save all trained variables.